+++ packages/sdp-custody/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "lib": ["ES2022", "DOM"]
+  },
   "include": ["src/**/*.ts"],
   "exclude": ["node_modules"]
 }
